requirements.txt
flask_basicauth
pandas
flask-bootstrap
dash
dash-renderer
dash-html-components
dash-core-components
dash-bootstrap-components
plotly
wget
epydemic
newsapi-python